Renamed dir 'flashcart' -> 'flashcarts', added tests for n64 flash carts
