+++ .gitignore
@@ -3,5 +3,5 @@
 /frontend/stats.json
 /frontend/tests/unit/rom-formats/data/psp/retail/*
 /frontend/tests/unit/rom-formats/data/psp/make-iso/*
-/frontend/tests/unit/save-formats/data/flashcart/goombaemulator/retail/*
+/frontend/tests/unit/save-formats/data/flashcarts/goombaemulator/retail/*
 /frontend/tests/config.local.js
