+++ frontend/tests/unit/save-formats/data/flashcarts/goombaemulator/Legend of Zelda, The - Link's Awakening (USA, Europe)-from-goomba.esv
